refactor(ui): standardize parameter naming and fix import paths
