+++ ui/src/app/(auth)/settings/page.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react";
 
 import { useTheme } from "@/contexts/ThemeContext";
-import { PasswordChangeCard } from "@/components/features/setting/PasswordChangeCard";
+import { PasswordChangeCard } from "@/components/features/settings/PasswordChangeCard";
 import { PageContainer } from "@/components/ui/PageContainer";
 import { PageHeader } from "@/components/ui/PageHeader";
 import { useAuth } from "@/contexts/AuthContext";
